+++ Changelog.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Generate Swagger schema and spec validation annotations.
 - Implement support for link relations in parser annotations.
 - Encode dashes (`-`) in URI template parameters.